Deleted refactored components.
